+++ tests/test_cli_py.py
@@ -8,13 +8,14 @@ from experitur.cli import clean, collect, do, run
 example_py = inspect.cleandoc(
     """
     from experitur import Experiment
+    import numpy as np
 
     @Experiment(
         parameters={
             "a1": [1],
             "a2": [2],
             "b": [1, 2],
-            "a": ["{a_{b}}"],
+            "a": ["{a{b}}"],
         })
     def baseline(trial):
         pass
@@ -34,6 +35,10 @@ example_py = inspect.cleandoc(
     @e3.command("cmd")
     def e3_cmd(trial):
         pass
+
+    @e3.command("experiment_cmd", target="experiment")
+    def e3_exp_cmd(experiment):
+        pass
     """
 )
 
@@ -44,7 +49,8 @@ def test_run():
         with open("example.py", "w") as f:
             f.write(example_py)
 
-        result = runner.invoke(run, ["example.py"], catch_exceptions=False)
+        result = runner.invoke(run, ["example.py"], catch_exceptions=True)
+        print(result.output)
         assert result.exit_code == 0
 
         result = runner.invoke(run, ["example.py", "--clean-failed"], input="y\n")
@@ -73,6 +79,30 @@ def test_do():
         result = runner.invoke(run, ["example.py"], catch_exceptions=False)
         assert result.exit_code == 0
 
+        # Assert that trial command succeedes
         result = runner.invoke(do, ["example.py", "cmd", "e3/_"])
         print(result.output)
         assert result.exit_code == 0
+
+        # Assert that experiment command succeedes
+        result = runner.invoke(do, ["example.py", "experiment_cmd", "e3"])
+        print(result.output)
+        assert result.exit_code == 0
+
+        # Assert that inexistent command fails
+        result = runner.invoke(
+            do, ["example.py", "inexistent_cmd", "e3/_"], catch_exceptions=False
+        )
+        print(result.output)
+
+        # Fail with click.UsageError
+        assert result.exit_code == 2
+
+        # Assert that inexistent trial fails
+        result = runner.invoke(
+            do, ["example.py", "cmd", "e3/inexistent_trial"], catch_exceptions=False
+        )
+        print(result.output)
+
+        # Fail with click.UsageError
+        assert result.exit_code == 2