+++ tests/test_merge_dicts.py
@@ -1,4 +1,5 @@
 from experitur.helpers.merge_dicts import merge_dicts
+import pytest
 
 
 def test_merge_dicts():
@@ -13,3 +14,9 @@ def test_merge_dicts():
     assert result is not a
 
     merge_dicts(a, b=None)
+
+    result = merge_dicts(a, list(b.items()), f=10)
+    assert result == {"a": 1, "b": {"c": 7, "d": 3}, "e": 4, "f": 10}
+
+    with pytest.raises(ValueError):
+        merge_dicts(a, 1)